refatora app.py
